requirements.txt

aiohttp==3.9.3
web3==6.10.0
python-dotenv==1.0.1
